moved club logo into folder
