+++ frontend/src/assets/cfaLogos/CFA Compact.png
