+++ frontend/src/assets/cfaLogos/CFA Wordmark.png
